[REF] core: merge test_testing_utilities with test_tests
