--- odoo/addons/test_testing_utilities/__manifest__.py
@@ -1,16 +1,0 @@
-{
-    'name': "Test testing utilities",
-    'category': "Hidden/Tests",
-    'description': """Non-trivial testing utilities can require models & all
-    
-This here module is useful to validate that they're doing what they're 
-supposed to do
-    """,
-    'depends': ['base', 'web'],
-    'data': [
-        'security/ir.model.access.csv',
-        'views/menu.xml',
-    ],
-    'author': 'Odoo S.A.',
-    'license': 'LGPL-3',
-}
