--- odoo/addons/test_testing_utilities/tests/__init__.py
@@ -1,8 +1,0 @@
-# -*- coding: utf-8 -*-
-from . import test_methods
-from . import test_lxml
-from . import test_form_impl
-from . import test_xml_tools
-from . import test_res_config
-from . import test_env
-from . import test_freeze_time
