+++ odoo/addons/test_tests/__manifest__.py
@@ -3,6 +3,10 @@
     'category': 'Hidden/Tests',
     'description': """Tests the Tests utilities.""",
     'depends': ['base', 'web'],
+    'data': [
+        'security/ir.model.access.csv',
+        'views/menu.xml',
+    ],
     'author': 'Odoo S.A.',
     'license': 'LGPL-3',
 }
