+++ odoo/addons/test_tests/security/ir.model.access.csv
@@ -29,5 +29,5 @@ access_ttu_product,access_ttu_product,model_ttu_product,base.group_user,1,0,0,0
 access_ttu_root,access_ttu_root,model_ttu_root,base.group_user,1,0,0,0
 access_ttu_child,access_ttu_child,model_ttu_child,base.group_user,1,0,0,0
 access_ttu_grandchild,access_ttu_grandchild,model_ttu_grandchild,base.group_user,1,0,0,0
-access_res_config_test,access_res_config_test,test_testing_utilities.model_res_config_test,base.group_user,1,0,0,0
+access_res_config_test,access_res_config_test,test_tests.model_res_config_test,base.group_user,1,0,0,0
 access_wide,access_wide,model_test_testing_utilities_wide,,0,0,0,0
