+++ odoo/addons/test_tests/tests/__init__.py
@@ -1,4 +1,9 @@
 from . import test_cases
+from . import test_form
+from . import test_freeze_time
+from . import test_methods
+from . import test_res_config
 from . import test_test_retry
 from . import test_test_suite
 from . import test_tests_tags
+from . import test_xml_tools
