+++ odoo/addons/test_tests/tests/test_xml_tools.py
