+++ odoo/addons/test_tests/views/menu.xml
@@ -80,7 +80,7 @@
             <form>
                 <field name="value"/>
                 <field name="v"/>
-                <field name="subs" context="{'list_view_ref': 'test_testing_utilities.editable_external'}"/>
+                <field name="subs" context="{'list_view_ref': 'test_tests.editable_external'}"/>
             </form>
         </field>
     </record>
@@ -243,7 +243,7 @@
                 <field name="subs"
                     invisible="value == 666"
                     context="{
-                        'list_view_ref': 'test_testing_utilities.o2m_modifier_ref_sub_tree',
+                        'list_view_ref': 'test_tests.o2m_modifier_ref_sub_tree',
                         'hide_c': True
                     }"/>
             </form>
