+++ odoo/addons/test_tools/tests/__init__.py
@@ -15,3 +15,4 @@ from . import test_safe_eval
 from . import test_set_expression
 from . import test_sql
 from . import test_translate
+from . import test_xml_utils
