+++ odoo/addons/test_tools/tests/test_xml_utils.py
